Move all dropq support functions to dropq_utils.py
