+++ taxcalc/dropq/dropq.py
@@ -9,19 +9,12 @@ from __future__ import print_function
 import time
 import numpy as np
 import pandas as pd
-from taxcalc.dropq.dropq_utils import (check_user_mods,
-                                       random_seed_from_subdict,
+from taxcalc.dropq.dropq_utils import (dropq_calculate,
                                        results,
-                                       drop_records)
-from taxcalc.dropq.dropq_utils import (create_dropq_distribution_table as
-                                       dropq_dist_table)
-from taxcalc.dropq.dropq_utils import (create_dropq_difference_table as
-                                       dropq_diff_table)
-from taxcalc.dropq.dropq_utils import create_json_table
-from taxcalc import (Calculator, Growfactors, Records,
-                     Policy, Consumption, Behavior, Growdiff,
-                     proportional_change_gdp, TABLE_LABELS,
-                     create_distribution_table)
+                                       random_seed,
+                                       dropq_summary,
+                                       create_json_table)
+from taxcalc import TABLE_LABELS, proportional_change_gdp
 
 
 # specify constants
@@ -44,248 +37,6 @@ GDP_ELAST_ROW_NAMES = ['gdp_elasticity']
 
 OGUSA_ROW_NAMES = ['GDP', 'Consumption', 'Investment', 'Hours Worked',
                    'Wages', 'Interest Rates', 'Total Taxes']
-
-
-def dropq_calculate(year_n, start_year,
-                    taxrec_df, user_mods,
-                    behavior_allowed, mask_computed):
-    """
-    The dropq_calculate function assumes specified user_mods is
-      a dictionary returned by the Calculator.read_json_parameter_files()
-      function with an extra key:value pair that is specified as
-      'gdp_elasticity': {'value': <float_value>}.
-    The function returns (calc1, calc2, mask) where
-      calc1 is pre-reform Calculator object calculated for year_n,
-      calc2 is post-reform Calculator object calculated for year_n, and
-      mask is boolean array if compute_mask=True or None otherwise
-    """
-    # pylint: disable=too-many-arguments,too-many-locals,too-many-statements
-
-    check_user_mods(user_mods)
-
-    # specify Consumption instance
-    consump = Consumption()
-    consump_assumptions = user_mods['consumption']
-    consump.update_consumption(consump_assumptions)
-
-    # specify growdiff_baseline and growdiff_response
-    growdiff_baseline = Growdiff()
-    growdiff_response = Growdiff()
-    growdiff_base_assumps = user_mods['growdiff_baseline']
-    growdiff_resp_assumps = user_mods['growdiff_response']
-    growdiff_baseline.update_growdiff(growdiff_base_assumps)
-    growdiff_response.update_growdiff(growdiff_resp_assumps)
-
-    # create pre-reform and post-reform Growfactors instances
-    growfactors_pre = Growfactors()
-    growdiff_baseline.apply_to(growfactors_pre)
-    growfactors_post = Growfactors()
-    growdiff_baseline.apply_to(growfactors_post)
-    growdiff_response.apply_to(growfactors_post)
-
-    # create pre-reform Calculator instance
-    recs1 = Records(data=taxrec_df.copy(deep=True),
-                    gfactors=growfactors_pre)
-    policy1 = Policy(gfactors=growfactors_pre)
-    calc1 = Calculator(policy=policy1, records=recs1, consumption=consump)
-    while calc1.current_year < start_year:
-        calc1.increment_year()
-    calc1.calc_all()
-    assert calc1.current_year == start_year
-
-    # optionally compute mask
-    if mask_computed:
-        # create pre-reform Calculator instance with extra income
-        recs1p = Records(data=taxrec_df.copy(deep=True),
-                         gfactors=growfactors_pre)
-        # add one dollar to total wages and salaries of each filing unit
-        recs1p.e00200 += 1.0  # pylint: disable=no-member
-        policy1p = Policy(gfactors=growfactors_pre)
-        # create Calculator with recs1p and calculate for start_year
-        calc1p = Calculator(policy=policy1p, records=recs1p,
-                            consumption=consump)
-        while calc1p.current_year < start_year:
-            calc1p.increment_year()
-        calc1p.calc_all()
-        assert calc1p.current_year == start_year
-        # compute mask that shows which of the calc1 and calc1p results differ
-        res1 = results(calc1)
-        res1p = results(calc1p)
-        mask = (res1.iitax != res1p.iitax)
-    else:
-        mask = None
-
-    # specify Behavior instance
-    behv = Behavior()
-    behavior_assumps = user_mods['behavior']
-    behv.update_behavior(behavior_assumps)
-
-    # always prevent both behavioral response and growdiff response
-    if behv.has_any_response() and growdiff_response.has_any_response():
-        msg = 'BOTH behavior AND growdiff_response HAVE RESPONSE'
-        raise ValueError(msg)
-
-    # optionally prevent behavioral response
-    if behv.has_any_response() and not behavior_allowed:
-        msg = 'A behavior RESPONSE IS NOT ALLOWED'
-        raise ValueError(msg)
-
-    # create post-reform Calculator instance
-    recs2 = Records(data=taxrec_df.copy(deep=True),
-                    gfactors=growfactors_post)
-    policy2 = Policy(gfactors=growfactors_post)
-    policy_reform = user_mods['policy']
-    policy2.implement_reform(policy_reform)
-    calc2 = Calculator(policy=policy2, records=recs2,
-                       consumption=consump, behavior=behv)
-    while calc2.current_year < start_year:
-        calc2.increment_year()
-    calc2.calc_all()
-    assert calc2.current_year == start_year
-
-    # increment Calculator objects for year_n years and calculate
-    for _ in range(0, year_n):
-        calc1.increment_year()
-        calc2.increment_year()
-    calc1.calc_all()
-    if calc2.behavior.has_response():
-        calc2 = Behavior.response(calc1, calc2)
-    else:
-        calc2.calc_all()
-
-    # return calculated Calculator objects and mask
-    return (calc1, calc2, mask)
-
-
-def random_seed(user_mods):
-    """
-    Compute random seed based on specified user_mods, which is a
-    dictionary returned by the Calculator.read_json_parameter_files()
-    function with an extra key:value pair that is specified as
-    'gdp_elasticity': {'value': <float_value>}.
-    """
-    ans = 0
-    for subdict_name in user_mods:
-        if subdict_name != 'gdp_elasticity':
-            ans += random_seed_from_subdict(user_mods[subdict_name])
-    return ans % np.iinfo(np.uint32).max  # pylint: disable=no-member
-
-
-def dropq_summary(df1, df2, mask):
-    """
-    df1 contains raw results for the standard plan X and X'
-    df2 contains raw results the user-specified plan (Plan Y)
-    mask is the boolean mask where X and X' match
-    """
-    # pylint: disable=too-many-locals
-
-    df1, df2 = drop_records(df1, df2, mask)
-
-    # Totals for diff between baseline and reform
-    dec_sum = (df2['tax_diff_dec'] * df2['s006']).sum()
-    bin_sum = (df2['tax_diff_bin'] * df2['s006']).sum()
-    pr_dec_sum = (df2['payrolltax_diff_dec'] * df2['s006']).sum()
-    pr_bin_sum = (df2['payrolltax_diff_bin'] * df2['s006']).sum()
-    combined_dec_sum = (df2['combined_diff_dec'] * df2['s006']).sum()
-    combined_bin_sum = (df2['combined_diff_bin'] * df2['s006']).sum()
-
-    # Totals for baseline
-    sum_baseline = (df1['iitax'] * df1['s006']).sum()
-    pr_sum_baseline = (df1['payrolltax'] * df1['s006']).sum()
-    combined_sum_baseline = (df1['combined'] * df1['s006']).sum()
-
-    # Totals for reform
-    sum_reform = (df2['iitax_dec'] * df2['s006']).sum()
-    pr_sum_reform = (df2['payrolltax_dec'] * df2['s006']).sum()
-    combined_sum_reform = (df2['combined_dec'] * df2['s006']).sum()
-
-    # Create difference tables, grouped by deciles and bins
-    diffs_dec = dropq_diff_table(df1, df2,
-                                 groupby='weighted_deciles',
-                                 res_col='tax_diff',
-                                 diff_col='iitax',
-                                 suffix='_dec', wsum=dec_sum)
-
-    diffs_bin = dropq_diff_table(df1, df2,
-                                 groupby='webapp_income_bins',
-                                 res_col='tax_diff',
-                                 diff_col='iitax',
-                                 suffix='_bin', wsum=bin_sum)
-
-    pr_diffs_dec = dropq_diff_table(df1, df2,
-                                    groupby='weighted_deciles',
-                                    res_col='payrolltax_diff',
-                                    diff_col='payrolltax',
-                                    suffix='_dec', wsum=pr_dec_sum)
-
-    pr_diffs_bin = dropq_diff_table(df1, df2,
-                                    groupby='webapp_income_bins',
-                                    res_col='payrolltax_diff',
-                                    diff_col='payrolltax',
-                                    suffix='_bin', wsum=pr_bin_sum)
-
-    comb_diffs_dec = dropq_diff_table(df1, df2,
-                                      groupby='weighted_deciles',
-                                      res_col='combined_diff',
-                                      diff_col='combined',
-                                      suffix='_dec', wsum=combined_dec_sum)
-
-    comb_diffs_bin = dropq_diff_table(df1, df2,
-                                      groupby='webapp_income_bins',
-                                      res_col='combined_diff',
-                                      diff_col='combined',
-                                      suffix='_bin', wsum=combined_bin_sum)
-
-    mX_dec = create_distribution_table(df1, groupby='weighted_deciles',
-                                       result_type='weighted_sum')
-
-    mY_dec = dropq_dist_table(df2, groupby='weighted_deciles',
-                              result_type='weighted_sum', suffix='_dec')
-
-    mX_bin = create_distribution_table(df1, groupby='webapp_income_bins',
-                                       result_type='weighted_sum')
-
-    mY_bin = dropq_dist_table(df2, groupby='webapp_income_bins',
-                              result_type='weighted_sum', suffix='_bin')
-
-    return (mY_dec, mX_dec, diffs_dec, pr_diffs_dec, comb_diffs_dec,
-            mY_bin, mX_bin, diffs_bin, pr_diffs_bin, comb_diffs_bin,
-            dec_sum, pr_dec_sum, combined_dec_sum,
-            sum_baseline, pr_sum_baseline, combined_sum_baseline,
-            sum_reform, pr_sum_reform, combined_sum_reform)
-
-
-def run_nth_year_gdp_elast_model(year_n, start_year,
-                                 taxrec_df, user_mods,
-                                 return_json=True):
-    """
-    The run_nth_year_gdp_elast_model function assumes user_mods is a
-    dictionary returned by the Calculator.read_json_parameter_files()
-    function with an extra key:value pair that is specified as
-    'gdp_elasticity': {'value': <float_value>}.
-    """
-    # create calc1 and calc2 calculated for year_n
-    (calc1, calc2, _) = dropq_calculate(year_n, start_year,
-                                        taxrec_df, user_mods,
-                                        behavior_allowed=False,
-                                        mask_computed=False)
-
-    # compute GDP effect given assumed gdp elasticity
-    gdp_elasticity = user_mods['gdp_elasticity']['value']
-    gdp_effect = proportional_change_gdp(calc1, calc2, gdp_elasticity)
-
-    # return gdp_effect results
-    if return_json:
-        gdp_df = pd.DataFrame(data=[gdp_effect], columns=['col0'])
-        gdp_elast_names_i = [x + '_' + str(year_n)
-                             for x in GDP_ELAST_ROW_NAMES]
-        gdp_elast_total = create_json_table(gdp_df,
-                                            row_names=gdp_elast_names_i,
-                                            num_decimals=5)
-        gdp_elast_total = dict((k, v[0]) for k, v in gdp_elast_total.items())
-        return gdp_elast_total
-    else:
-        return gdp_effect
 
 
 def run_nth_year_tax_calc_model(year_n, start_year,
@@ -411,6 +162,39 @@ def run_nth_year_tax_calc_model(year_n, start_year,
             cdf_dec_table_i, mY_bin_table_i, mX_bin_table_i, df_bin_table_i,
             pdf_bin_table_i, cdf_bin_table_i, fiscal_yr_total_df,
             fiscal_yr_total_bl, fiscal_yr_total_rf)
+
+
+def run_nth_year_gdp_elast_model(year_n, start_year,
+                                 taxrec_df, user_mods,
+                                 return_json=True):
+    """
+    The run_nth_year_gdp_elast_model function assumes user_mods is a
+    dictionary returned by the Calculator.read_json_parameter_files()
+    function with an extra key:value pair that is specified as
+    'gdp_elasticity': {'value': <float_value>}.
+    """
+    # create calc1 and calc2 calculated for year_n
+    (calc1, calc2, _) = dropq_calculate(year_n, start_year,
+                                        taxrec_df, user_mods,
+                                        behavior_allowed=False,
+                                        mask_computed=False)
+
+    # compute GDP effect given assumed gdp elasticity
+    gdp_elasticity = user_mods['gdp_elasticity']['value']
+    gdp_effect = proportional_change_gdp(calc1, calc2, gdp_elasticity)
+
+    # return gdp_effect results
+    if return_json:
+        gdp_df = pd.DataFrame(data=[gdp_effect], columns=['col0'])
+        gdp_elast_names_i = [x + '_' + str(year_n)
+                             for x in GDP_ELAST_ROW_NAMES]
+        gdp_elast_total = create_json_table(gdp_df,
+                                            row_names=gdp_elast_names_i,
+                                            num_decimals=5)
+        gdp_elast_total = dict((k, v[0]) for k, v in gdp_elast_total.items())
+        return gdp_elast_total
+    else:
+        return gdp_effect
 
 
 def format_macro_results(diff_data, return_json=True):
