+++ taxcalc/tests/test_dropq.py
@@ -1,20 +1,21 @@
-# pylint: disable=missing-docstring,wildcard-import,unused-wildcard-import
+# pylint: disable=missing-docstring,undefined-variable
+# pylint: disable=wildcard-import,unused-wildcard-import,import-error
 import os
 import numpy as np
 import pandas as pd
 import pytest
-from taxcalc.dropq.dropq_utils import *  # pylint disable=wildcard-import
-from taxcalc.dropq import *  # pylint disable=wildcard-import
+from taxcalc.dropq.dropq_utils import *
+from taxcalc.dropq import *
 from taxcalc import Policy, Records, Calculator
 from taxcalc import multiyear_diagnostic_table
 
 
 USER_MODS = {
     'policy': {
-        2016: {'_II_rt4': [0.39, 0.40, 0.41],
-               '_PT_rt4': [0.39, 0.40, 0.41],
-               '_II_rt3': [0.31, 0.32, 0.33],
-               '_PT_rt3': [0.31, 0.32, 0.33]}
+        2016: {'_II_rt3': [0.33],
+               '_PT_rt3': [0.33],
+               '_II_rt4': [0.33],
+               '_PT_rt4': [0.33]}
     },
     'consumption': {
         2016: {'_MPC_e20400': [0.01]}
@@ -38,22 +39,25 @@ def puf_path(tests_path):
     """
     return os.path.join(tests_path, '..', '..', 'puf.csv')
 
-
+@pytest.mark.one
 def test_check_user_mods_errors():
-    user_mods = list()
+    check_user_mods(USER_MODS)
+    seed1 = random_seed(USER_MODS)
     with pytest.raises(ValueError):
-        check_user_mods(user_mods)
+        check_user_mods(list())
     usermods = USER_MODS
-    usermods.pop('growdiff_baseline')
+    behavior_subdict = usermods.pop('behavior')
     with pytest.raises(ValueError):
         check_user_mods(usermods)
-    usermods['growdiff_baseline'] = dict()
+    usermods['behavior'] = behavior_subdict
     usermods['unknown_key'] = dict()
     with pytest.raises(ValueError):
         check_user_mods(usermods)
     usermods.pop('unknown_key')
+    seed2 = random_seed(usermods)
+    assert seed1 == seed2
 
-
+@pytest.mark.one
 def test_run_nth_year_value_errors(puf_1991_path):
     recs = pd.read_csv(puf_1991_path)
     usermods = USER_MODS
@@ -64,7 +68,7 @@ def test_run_nth_year_value_errors(puf_1991_path):
     with pytest.raises(ValueError):
         run_nth_year_gdp_elast_model(1, 2013, recs, usermods, False)
 
-
+@pytest.mark.one
 @pytest.mark.parametrize('resjson', [True, False])
 def test_run_tax_calc_model(puf_1991_path, resjson):
     recs = pd.read_csv(puf_1991_path)
@@ -78,7 +82,7 @@ def test_run_tax_calc_model(puf_1991_path, resjson):
         else:
             assert isinstance(res[idx], pd.DataFrame)
 
-
+@pytest.mark.one
 @pytest.mark.parametrize('resjson', [True, False])
 def test_run_gdp_elast_model(puf_1991_path, resjson):
     usermods = USER_MODS
@@ -92,7 +96,7 @@ def test_run_gdp_elast_model(puf_1991_path, resjson):
     else:
         assert isinstance(res, float)
 
-
+@pytest.mark.one
 def test_format_macro_results():
     data = [[1.875e-03, 1.960e-03, 2.069e-03, 2.131e-03, 2.179e-03, 2.226e-03,
              2.277e-03, 2.324e-03, 2.375e-03, 2.426e-03, 2.184e-03, 2.806e-03],
@@ -138,7 +142,7 @@ def test_format_macro_results():
     res_df = format_macro_results(data, return_json=False)
     assert res_df.equals(pd.DataFrame(data))
 
-
+@pytest.mark.one
 def test_random_seed_from_subdict():
     """
     Test except logic in try statement in random_seed_from_subdict function.
@@ -152,7 +156,7 @@ def test_random_seed_from_subdict():
     seed2 = random_seed_from_subdict(dct)
     assert seed1 == seed2
 
-
+@pytest.mark.one
 def test_chooser_error():
     dframe = pd.DataFrame(data=[[0, 1], [0, 2], [0, 3],
                                 [0, 4], [0, 5], [0, 6],
@@ -165,13 +169,13 @@ def test_chooser_error():
     with pytest.raises(ValueError):
         chooser(dframe['zeros'])
 
-
+@pytest.mark.one
 def test_format_print_error():
     arr = np.array([1], dtype='i2')
     with pytest.raises(NotImplementedError):
         format_print(arr[0], arr.dtype, 2)
 
-
+@pytest.mark.one
 def test_create_json_table():
     dframe = pd.DataFrame(data=[[1., 2, 3], [4, 5, 6], [7, 8, 9]],
                           columns=['a', 'b', 'c'])
@@ -181,7 +185,7 @@ def test_create_json_table():
            '2': ['7.00', '8', '9']}
     assert ans == exp
 
-
+@pytest.mark.one
 @pytest.mark.parametrize('groupby, result_type',
                          [('small_income_bins', 'weighted_sum'),
                           ('large_income_bins', 'weighted_sum'),
@@ -197,13 +201,11 @@ def test_dropq_dist_table(groupby, result_type, puf_1991_path):
     (res, _) = drop_records(res, res, mask)
     if groupby == 'other_income_bins' or result_type == 'other_avg':
         with pytest.raises(ValueError):
-            create_dropq_distribution_table(res, groupby=groupby,
-                                            result_type=result_type,
-                                            suffix='_bin')
+            dropq_dist_table(res, groupby=groupby,
+                             result_type=result_type, suffix='_bin')
     else:
-        create_dropq_distribution_table(res, groupby=groupby,
-                                        result_type=result_type,
-                                        suffix='_bin')
+        dropq_dist_table(res, groupby=groupby,
+                         result_type=result_type, suffix='_bin')
 
 @pytest.mark.one
 @pytest.mark.parametrize('groupby, res_column',
@@ -229,15 +231,15 @@ def test_dropq_diff_table(groupby, res_column, puf_1991_path):
     dec_sum = (res2['tax_diff_dec'] * res2['s006']).sum()
     if groupby == 'other_deciles':
         with pytest.raises(ValueError):
-            create_dropq_difference_table(res1, res2, groupby=groupby,
-                                          res_col=res_column, diff_col='iitax',
-                                          suffix='_dec', wsum=dec_sum)
+            dropq_diff_table(res1, res2, groupby=groupby,
+                             res_col=res_column, diff_col='iitax',
+                             suffix='_dec', wsum=dec_sum)
     else:
-        create_dropq_difference_table(res1, res2, groupby=groupby,
-                                      res_col=res_column, diff_col='iitax',
-                                      suffix='_dec', wsum=dec_sum)
+        dropq_diff_table(res1, res2, groupby=groupby,
+                         res_col=res_column, diff_col='iitax',
+                         suffix='_dec', wsum=dec_sum)
 
-
+@pytest.mark.one
 @pytest.mark.requires_pufcsv
 def test_with_pufcsv(puf_path):  # pylint: disable=redefined-outer-name
     # pylint: disable=too-many-locals
@@ -245,10 +247,10 @@ def test_with_pufcsv(puf_path):  # pylint: disable=redefined-outer-name
     start_year = 2016
     reform_year = start_year + 1
     reforms = dict()
-    reforms['_II_rt4'] = [0.39, 0.40, 0.41]
-    reforms['_PT_rt4'] = [0.39, 0.40, 0.41]
-    reforms['_II_rt3'] = [0.31, 0.32, 0.33]
-    reforms['_PT_rt3'] = [0.31, 0.32, 0.33]
+    reforms['_II_rt3'] = [0.33]
+    reforms['_PT_rt3'] = [0.33]
+    reforms['_II_rt4'] = [0.33]
+    reforms['_PT_rt4'] = [0.33]
     usermods = dict()
     usermods['policy'] = {reform_year: reforms}
     usermods['consumption'] = {}
@@ -282,4 +284,9 @@ def test_with_pufcsv(puf_path):  # pylint: disable=redefined-outer-name
     dropq_reform_revenue *= 1e-9  # convert to billions of dollars
     diff = abs(fulls_reform_revenue - dropq_reform_revenue)
     # assert that dropq revenue is similar to the fullsample calculation
-    assert diff / fulls_reform_revenue < 0.01
+    proportional_diff = diff / fulls_reform_revenue
+    frmt = 'f,d,adiff,pdiff=  {:.4f}  {:.4f}  {:.4f}  {}'
+    print(frmt.format(fulls_reform_revenue, dropq_reform_revenue,
+                      diff, proportional_diff)) 
+    assert proportional_diff < 0.001  # one-tenth of one percent
+    assert 1 == 2
